+++ contracts/metadata/CoreMetadata.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 
-pragma solidity ^0.6.8;
+pragma solidity 0.6.8;
 
 import "@openzeppelin/contracts/introspection/ERC165.sol";
 import "@animoca/ethereum-contracts-core_library/contracts/utils/types/UInt256Extract.sol";