+++ contracts/mocks/token/ERC20/ERC20WithOperatorsMock.sol
@@ -1,6 +1,6 @@
 
 // SPDX-License-Identifier: MIT
 
-pragma solidity ^0.6.8;
+pragma solidity 0.6.8;
 
 import "@animoca/ethereum-contracts-erc20_base/contracts/mocks/token/ERC20/ERC20WithOperatorsMock.sol";