+++ contracts/token/ERC1155/ERC1155TokenReceiver.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 
-pragma solidity ^0.6.8;
+pragma solidity 0.6.8;
 
 import "@openzeppelin/contracts/introspection/ERC165.sol";
 import "./IERC1155TokenReceiver.sol";