+++ src/constants.js
@@ -17,7 +17,7 @@ const ERC721Enumerable_InterfaceId = interfaces.ERC721Enumerable.id;
 const ERC721Exists_InterfaceId_Experimental = interfaces.ERC721Exists_Experimental.id;
 const ERC721Receiver_InterfaceId = interfaces.ERC721Receiver;
 const ERC1155_InterfaceId = interfaces.ERC1155.id;
-const ERC1155Collections_InterfaceId_Experimental = interfaces.ERC1155Collections_Experimental.id;
+const ERC1155AssetCollections_InterfaceId_Experimental = interfaces.ERC1155AssetCollections_Experimental.id;
 const ERC1155MetadataURI_InterfaceId = interfaces.ERC1155MetadataURI.id;
 const ERC1155TokenReceiver_InterfaceId = interfaces.ERC1155TokenReceiver;
 
@@ -43,11 +43,11 @@ module.exports = {
     // Bytes4
     ERC721Received_MagicValue, ERC1155Received_MagicValue, ERC1155BatchReceived_MagicValue,
     ERC721_InterfaceId, ERC721Metadata_InterfaceId, ERC721Enumerable_InterfaceId, ERC721Exists_InterfaceId_Experimental, ERC721Receiver_InterfaceId,
-    ERC1155_InterfaceId, ERC1155MetadataURI_InterfaceId, ERC1155Collections_InterfaceId_Experimental, ERC1155TokenReceiver_InterfaceId,
+    ERC1155_InterfaceId, ERC1155MetadataURI_InterfaceId, ERC1155AssetCollections_InterfaceId_Experimental, ERC1155TokenReceiver_InterfaceId,
     Bytes4: {
         ERC721Received_MagicValue, ERC1155Received_MagicValue, ERC1155BatchReceived_MagicValue,
         ERC721_InterfaceId, ERC721Metadata_InterfaceId, ERC721Enumerable_InterfaceId, ERC721Exists_InterfaceId_Experimental, ERC721Receiver_InterfaceId,
-        ERC1155_InterfaceId, ERC1155MetadataURI_InterfaceId, ERC1155Collections_InterfaceId_Experimental, ERC1155TokenReceiver_InterfaceId,
+        ERC1155_InterfaceId, ERC1155MetadataURI_InterfaceId, ERC1155AssetCollections_InterfaceId_Experimental, ERC1155TokenReceiver_InterfaceId,
     },
 
     // Bits Layouts