+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 ## 3.0.0 (20/05/2020)
 
 ### Breaking changes
- * Updated compiler to `solc:0.6.8`.
- * Updated dependecy to `@animoca/ethereum-contracts-core_library:2.0.0`.
- * Updated dependecy to `@animoca/ethereum-contracts-erc20_base:2.0.0`.
+ * Updated compiler to `solc:0.6.8` and fixed solidity version in contract filed.
+ * Updated `@animoca/ethereum-contracts-core_library` to version 3 and downgraded it to be a dev dependency.
+ * Updated `@animoca/ethereum-contracts-erc20_base` to version 3 and downgraded it to be a dev dependency.
  * Added non-ERC721 versions of the inventory contracts.
  * Major refactor of contracts and tests, increased tests coverage.